third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-on-object.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#obj {
  width: 50px;
  height: 50px;
  padding: 50px;
}
</style>

<h1>Verifies that mouse activities on an object fire pointerevents</h1>

<object id="obj"></object>

<div id="log"></div>

<script>
var receivedEvents = [];

function testReceivedEvents(expectedEvents, testLabel) {
  test(function() {
    assert_array_equals(receivedEvents, expectedEvents);
  }, testLabel);
  receivedEvents = [];
}

function init() {
  ["mousedown", "mouseup", "mousemove", "pointerdown", "pointerup", "pointermove"].forEach(function(eventName) {
    document.getElementById("obj").addEventListener(eventName, function(event) {
      receivedEvents.push(event.type);
    });
  });
}

function runTests() {
  var rect = document.getElementById("obj").getBoundingClientRect();

  eventSender.mouseMoveTo(rect.left + 5, rect.top + 5);
  eventSender.mouseDown(0);
  eventSender.mouseUp(0);
  testReceivedEvents(["pointermove", "mousemove", "pointerdown", "mousedown", "pointerup", "mouseup"], "Click on object");

  eventSender.mouseDown(0);
  eventSender.mouseMoveTo(rect.left - 5, rect.top -5);
  eventSender.mouseUp(0);
  testReceivedEvents(["pointerdown", "mousedown", "pointermove", "mousemove", "pointerup", "mouseup"], "Drag from object");
}

test(function() {
  if (!window.eventSender)
    assert_true(true, "No eventSender, skipped tests");
  else {
    init();
    runTests();
  }

  done();
}, "Verifies that mouse activities on an object fire pointerevents");
</script>
